[camera] Support pausing and resuming video recording
