+++ apps/native-component-list/src/navigation/ExpoComponentsStackNavigator.tsx
@@ -8,6 +8,7 @@ import { optionalRequire } from './routeBuilder';
 import { TabBackground } from '../components/TabBackground';
 import TabIcon from '../components/TabIcon';
 import { Layout } from '../constants';
+import { CameraScreens } from '../screens/Camera/CameraScreen';
 import ExpoComponents from '../screens/ExpoComponentsScreen';
 import { ImageScreens } from '../screens/Image/ImageScreen';
 import { VideoScreens } from '../screens/Video/VideoScreen';
@@ -52,18 +53,6 @@ export const Screens: ScreenConfig[] = [
       return optionalRequire(() => require('../screens/Camera/CameraScreen'));
     },
     name: 'Camera',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Camera/CameraScreenBarcode'));
-    },
-    name: 'Camera (barcode)',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Camera/CameraScreenBarcodeFromURL'));
-    },
-    name: 'Camera (barcode from URL)',
   },
   {
     getComponent() {
@@ -436,6 +425,7 @@ export const Screens: ScreenConfig[] = [
     },
     name: 'MeshGradient',
   },
+  ...CameraScreens,
   ...ImageScreens,
   ...VideoScreens,
 ];
