+++ packages/expo-camera/android/src/main/java/expo/modules/camera/ExpoCameraView.kt
@@ -126,6 +126,7 @@ class ExpoCameraView(
   private var recorder: Recorder? = null
   private var barcodeFormats: List<BarcodeType> = emptyList()
   private var glSurface: SurfaceTexture? = null
+  private var isRecording = false
 
   private var previewView = PreviewView(context).apply {
     elevation = 0f
@@ -334,6 +335,7 @@ class ExpoCameraView(
           }
         }
         .start(ContextCompat.getMainExecutor(context)) { event ->
+          isRecording = true
           when (event) {
             is VideoRecordEvent.Finalize -> {
               when (event.error) {
@@ -363,6 +365,21 @@ class ExpoCameraView(
         "Starting video recording failed - could not create video file.",
         null
       )
+  }
+
+  fun stopRecording() {
+    isRecording = false
+    activeRecording?.close()
+  }
+
+  fun toggleRecording() {
+    activeRecording?.let {
+      if (isRecording) {
+        it.pause()
+      } else {
+        it.resume()
+      }
+    }
   }
 
   @SuppressLint("UnsafeOptInUsageError")
