+++ packages/expo-camera/src/CameraView.tsx
@@ -228,6 +228,13 @@ export default class CameraView extends Component<CameraViewProps> {
   }
 
   /**
+   * Pauses or resumes the video recording. Only has an effect if there is an active recording.
+   */
+  async toggleRecordingAsync() {
+    return this._cameraRef.current?.toggleRecording();
+  }
+
+  /**
    * Stops recording if any is in progress.
    */
   stopRecording() {
